Nydusify: preferred use of existing oci manifest
If the multi-platform option is enabled, Nydusify will try to merge the OCI
manifest with the Nydus manifest into a manifest list, but this may cause
the existing OCI manifest on the remote to be overwritten, and the OCI
image's digest be changed, then the overwritten image may be deleted by
registry garbage collection, which may cause problems for users.

The fix attempts to use the remote OCI manifest first and the source image's
OCI manifest second when merging manifests.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>