Nydusify: fix lost blob layers in manifest
For the registry backend, we need to write the blob layer to the Nydus image
manifest to prevent it from being deleted by the registry's garbage collection.

Add this use case to the checker to compare the blob list differences between
bootstrap layer annotation, manifest layers and blob table in bootstrap.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>