Builder: add an index to chunk info
Add an index to OndiskChunkInfo struct, it's allocated sequentially
starting from 0 for one blob. In the blobcache runtime, use this
index to query the cache ready state of a chunk.

Add an chunk_count to the entry of OndiskBlobTable, it records the
number of chunks in one blob. The former readahead_offset field is
deprecated and the chunk_count field takes its place, the readahead
offset should always zero because the data segment to be prefetched
always be put in the head of blob file.

In addition, a blob index field has been added to the entry in the
blob table to facilitate indexing the blob at blobcache runtime.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>