storage/dummycache: don't return error in prefetch() and stop_prefetch()
There's no need for dummycache to do prefetch() nor stop_prefetch().
But currently we return StorageError::Unsupported, and this results in a
error message everytime we switch a localfs backend rafs when stopping
prefetch, e.g. from blob file on NAS to local fs.

  "ERRO Unsupported"

So just return OK in prefetch() and stop_prefetch().

Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>