Blobcache: add ut for corrupted chunkmap
This test deletes the blobcache file without deleting its
corresponding chunkmap file. The expected result is that
newly created blobcache instance will read corrupt blob
data, but old blobcache instance should not be affected,
and reading blob data should not be affected if only the
chunkmap file is deleted but not the blobcache file.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>