rafs: use mtime instead of ctime
The modified timestamp (mtime) signifies the last time the contents
of a file were modified. A program or process either edited or
manipulated the file. “Modified” means something inside the file was
amended or deleted, or new data was added.

The changed timestamps aren’t referring to changes made to the contents
of a file. Rather, it’s the time at which the metadata related to the
file was changed. File permission changes, for example, will update the
changed timestamp.

So the rafs metadata should record `mtime` instead of `ctime` because
`mtime` is often used to detect content change.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>