nydus-image: fix disorder of writing chunks to blob
In rafs disk layout, we should enforce a rule that chunks
are dumped/written to blob file according to order of inode
number.

When prefetch files list is specified, e.g. "/", the specified
files will be dumped before dumping other files. But the `readahead_files`
collection is not sorted by its key - inode index.

So fix this by sorting it before consuming it.

This problem might affect prefetch merging possibility.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>