Builder: implement diff build
Diff build is a new multi-layers image build workflow that allows
multiple directory paths to be passed into the builder at once as
arguments, e.g.

```
nydus-image create ... --source-type diff \
/path/to/snapshot.1 /path/to/snapshot.2 ...  /path/to/snapshot.N
```

or

nydus-image create ... --source-type diff --diff-overlay-hint \
/path/to/snapshot.1 /path/to/snapshot.2 ...  /path/to/snapshot.N \
/path/to/upper.1 /path/to/upper.2 ... /path/to/upper.N
```

Where snapshot directories are the committed snapshots of an image
in containerd, sorted by layer, each snapshot is the merge (overlayed)
dir of all lower snapshots below it. And where upper directories (only
need be specified when --diff-overlay-hint option is enabled) are
the upper directories of each snapshot for overlayfs, they are used
to tell diff build to find added or modified files in each layer
faster. The diff build workflow is divided into two phases:

Phase 1: walk snapshot.n and snapshot.n+1, compare the differences
in regular files between the dirs, dump the chunk of file data added
or modified in snapshot.n+1 to a blob, which is the nydus blob of
snapshot.n+1, when --diff-overlay-hint option (only for overlayfs)
is enabled, added or modified files can be found faster from the
upper.n+1 directory, thus speeding up the blob dumping. In this way,
iteratively build each layer of dir in sequence.

Phase 2: walk the last snapshot directory, dumping the file metadata
to a bootstrap file, which is the nydus bootstrap of the entire image.

The advantage of diff build is that it can build the blob of each
layer concurrently in phase 1, which greatly improves the speed of
building multi-layers image and eliminates the need to consider the
processing of whiteouts in upper layer.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>