nydus-image: add support to generate erofs image
This enables nydus-image to generate an erofs image when 'v6-mode'
option is on.

The idea is to reuse v5 inode info and v5 chunk info to generate an
erofs image, so for now the only difference between v5 and v6 is the
layout of bootstrap while the layour of blob file remains the same.

Note that v6-mode also forces to set 'aligned-chunk' so that erofs
kerne fs can mount it directly.

Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>