Backend: fix registry authorization
On harbor or ghcr registry implementation, token service only accept
HTTP GET request and must put auth string to the header.

The information needed for getting token needs to be placed both in
the query and in the body to be compatible with different registry
implementations, which have been tested on these platforms:
docker hub, harbor, github ghcr, aliyun acr.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>