nydus-image: add s_devt_slotoff for devtable of rafs v6
Erofs in kernel needs a devt_slotoff and the affiliated device table
to map the correct one in its multi-device mode.

This fills the above gap when buidling a rafs v6 image.

This also aligns bootstrap size to 4k so that it can be read correctly
by erofs over loop device.

Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>