nydusd: recover option --shared-dir with feature fusedev
Even we exclude this option from nydusd with feature fusedev applied,
it can still be used from API mount.
Like libfuse, we just want to provide a simple example that built-in
fuse-rs works well and test can be composed to verify code in nydusd/rafs
when linking to virtiofs front-end.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>