Builder: introduce artifact storage
Add new ArtifactBufferWriter struct, it provides a writer to allow
writing bootstrap or blob data to a single file or in a directory.

It's useful in diff build scenario, we need to dump bootstrap for
every layer concurrently, with the aid of BootstrapManager, we can
create multiple BootstrapContexts to write bootstrap to a directory.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>