storage: fix uncompressed_size in BlobMetaState
Commit b434a3e4c4 ("Storage/meta: add munmap for mmaped range")
introduced the regression.

Since the name is quite confusing, this is to clarify it a little bit,
BlobMetaState's [un]compressed_size is used by
get_chunk_[un]compressed() and is considered as file length of a blob,
NOT a blob's ondisk chunk info's size.

Given the above, we are not able to get correct unmap length from
uncompressed_size, this adds a 'unmap_len' for Drop.

Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>