nydus-image: fix a regression of compact inode
A regression was introduced in commit 0a8863b5 ("rafs: support 32-byte
compact inodes") and it is about the inconsistence that different
types of inode are used in the build stage and the dump stage.
This may end up with corrupted bootstrap.

We need to decide to use compact inode since the build stage instead
of the dump stage.

Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>