Builder: some fixups for diff build cache
Add the `ordered_blobs` field in outout JSON file:

For buildkit scenario, we need to know which layer has no blob output,
so that we can skip the handle of build cache, adding the `ordered_blobs`
for giving the information.

The chunk of upper snapshot should be added to correct position in CachedNodes:

In previous, we use CachedNodes to make the final bootstrap of image refer
to the modified files in upper snapshot, not the files in lower snapshot,
but we didn't fill None in CachedNodes for the lower snapshot that does not
contain the file, this causes that generating some chunks included incorrect
blob index in final bootstrap.

For debug requirement of diff build, this patch disable some chaotic logs,
and remove some unnecessary v6 function calls for v5 build.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>