Nydusify: relax blob list check in checker
Issue: in previous, we put the blob list encoded in JSON string to
the label of bootstrap layer in nydus manifest, but this may be causes
that over 4k size limitation in containerd content store when a nydus
image has many blob layers.

Fix: for compatibility, still, keep the blob list in bootstrap layer
when use `nydusify convert` command, but remove the blob list consistency
check in checker for the new buildkit scenario.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>