Builder: fix blob race in diff build
Issue: in diff build, blob is generated in parallel, maybe have same
blob generated by different layers if no `--chunk-dict` option be enabled
to dedup chunk, in this time, the release method in ArtifactBufferWriter
will remove a blob (before rename) that be about to open in buildkit, this
causes that failure in open with error: "open: not found".

Fix: don't cleanup existing blob file, the newly generated blob is ignored,
it doesn't matter.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>