nydus-image: support chunk-dict for diff hint builder
The diff hint build has mapped the layer to the order of the blob
table, in order to avoid breaking diff hint build, we need to append
blobs of chunk-dict to the end of the blob table after the build
finishes. Meanwhile, we also need to set correct index of blobs of
chunk-dict in the blob table so that dup'd chunk can have the correct
chunk.blob_index on the go.

We do them separately, since the blob table is determined at the
beginning of diff-hint build, we are able to set the correct index for
blobs of chunk-dict before the build starts, and when the build
finishes, we then append the blobs of chunk-dict to to the blob table.

Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>
Signed-off-by: Yan Song <imeoer@linux.alibaba.com>