bugfix: storage: Fix the race problem when decompressing with lz4
Lz4 does not support concurrent decompression of the same data into
the same piece of memory. There will be multiple containers mmap the
same file, causing the buffer to be shared between different
processes. This will cause data errors due to race issues when
decompressing with lz4. We solve this problem by creating a temporary
memory to hold the decompressed data.

Because this process will only be executed when the blob.meta file is
created for the first time, which means that a machine will only
execute the process once when the blob.meta is created for the first
time, the memory consumption and performance impact are relatively
small.

Signed-off-by: gexuyang <gexuyang@linux.alibaba.com>