nydus-image: Bug fix for the `..` disappear in some subdirs of the root directory
This bug is due to we use 0 as the root directory's nid.
The glibc's getdents/readdirs wrapper treat file with ino 0 as
deleted(https://code.woboq.org/userspace/glibc/sysdeps/posix/readdir.c.html),
so we need to avoid using 0 as root inode's nid.
To do this, we make the superblock's meta_blkaddr one block ahead of the inode table

fix #403

Signed-off-by: Qi Wang <mpiglet@outlook.com>