builder: support tar like build
The `inline-bootstrap` option merges the blob and bootstrap into one
file. We need some header to index the location of the blob and bootstrap,
this patch uses a tar header structure that arranges the data as follows.

blob_data | blob_tar_header | bootstrap_data | boostrap_tar_header

This is a tar-like structure, except that we put the tar header after the
data. The advantage is that we do not need to determine the size of the data
first, so that we can write the blob data by stream without seek to improve
the performance of the blob dump by using fifo, if we need to read the bootstrap
data quickly, first need to read the 512 bytes tar header from the end of blob
file first, and then seek offset to read bootstrap data.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>