nydus-image: Feature, add a block allocator for reducing bootstrap size by reusing used blocks
Firstly, we add a `dirents_offset` field in `struct Node`, which is used by directory files.
Previously, the dirents have to be placed the next block where its dir inode was placed, by
introducing this field, we allow the dirents to be seperated from its dir inode.
Secondly, we add an `available_blocks` field in `struct BootstrapCtx`,
to record blocks that are used but still contain unused space.

feature #348

Signed-off-by: Qi Wang <wangqi@linux.alibaba.com>