builder: fix lost hardlink
The merge operation in builder (nydus-image) use `RafsSuper::walk_inodes`
method to walkthrough the whole filesystem inodes by DFS order, but it
does not call the callback for the hardlink pair, which causes some hardlink
files to be lost in the final bootstrap after merge.

The root cause is that the `RafsSuper::get_inode` method always gets only
the RafsInode of a particular file name in the hardlink pair because the ino
numbers of the hardlink pair are the same.

Fix by passing `RafsInode` trait object in `RafsSuper::walk_inodes` directly.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>