rafs: delete RafsSuperBlobs trait
method get_blobs of RafsSuperBlobs has the same signature as
get_blob_infos of RafsSuperBlock,
and RafsSuperBlock is inheritance of RafsSuperBlobs.

But the only one use of the get_blobs is based on type of
RafsSuperBlock, so RafsSuperBlobs is no meanings, it can
be deleted and replaced by get_blob_infos method.

Signed-off-by: bin liu <bin@hyper.sh>