rafs v6: bug fix for the incorrect dirents process
The dir file's dirents may occupy multiple blocks, and only the last
block may not be fully used. Previously, the last dirent of each block
is treated as the last dirent of the last block, which causes the
nydusd-fusedev to panic when using rafs v6 format.

To fix this bug, we process the last dirent of the last block different
from the last dirent of the other blocks.

fix #428

Signed-off-by: Qi Wang <wangqi@linux.alibaba.com>