nydus-image: fix invalid meta_addr in rafs v6
When building a rafs v6 image against a directory with only empty
files, it turns out that meta_addr is 0 which overlaps with the
superblock area.

This fixes the problem by setting meta_addr to EROFS_BLOCK_SIZE in the
above case.

This also makes sure that the root inode starts with RAFS_ROOT_INODE
to keep consistent with rafs v5.

Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>