storage: clean data prefetch logic
The data prefetch related code has been refactored several times,
and it becomes a little complex. So go over code again to remove some
legacy code.

The most important change is to use reference count to track prefetch
status instead of an on/off flag. It gets messy with the on/off flag
when dealing with blob sharing/reuse.

Signed-off-by: Jiang Liu <gerry@linux.alibaba.com>