nydus-image: fixed a bug caused by empty blob and prefetched combined scenes
When building a layer independently, if the layer does not produce any blobs, but the prefetch is set,
the position of the prefetch table and the position of the inode table will overlap,
resulting in the prefetch table overwriting part of the inode table.

blob table|prefetch table|inode table

blob table + prefetch table is a whole, 4K alignment, if blob table size is equal to 0, but prefetchtable is not equal to 0,
will cause the prefetch table to exist, but will not be 4k aligned, and overlap with the inode table.

Related code

        let orig_meta_addr = bootstrap_ctx.nodes[0].offset - EROFS_BLOCK_SIZE;
        let meta_addr = if blob_table_size > 0 {
            align_offset(
                blob_table_offset + blob_table_size + prefetch_table_size as u64,
                EROFS_BLOCK_SIZE as u64,
            )
        } else {
            orig_meta_addr
        };
        let root_nid = calculate_nid(
            bootstrap_ctx.nodes[0].offset - orig_meta_addr + meta_addr,
            meta_addr,
        );
Signed-off-by: zyfjeff <zyfjeff@linux.alibaba.com>