nydusd: make copen always return right file size
If we fail to create a blob object and return an error in the copen cmd
for blob fscache cookie , the erofs will be mounted but the blob data
can not be accessed. The following requests will not be sent to the user
daemon , then it has no chance to handle this case.
 
Make copen always return the right file size , then the following
requests may be dispatched to user daemon, we can do some retry later.

For now, use an asynchronous thread to create blob object on open cmd,
and will wait for the result before first ondemand read.

Signed-off-by: Xin Yin <yinxin.x@bytedance.com>