storage: support configure proxy use_http
Dragonfly supports configuration `useHTTPS` while
work a http proxy. In this mode, nydusd can access
dfdaemon via http while dfdaemon is accessing registry
using https. So CA is not needed to be passed to dfdaemon.

The limit is that registry scheme must be set to http ending
up with reading original registry failure when fallback.

With configuring proxy with use_http, it will replace the
scheme when sending requests to dfdaemon.

Signed-off-by: Changwei Ge <gechangwei@bytedance.com>