rafs: fix a bug in clone ArcSwap object
The documentation from ArcSwap states:
Previous version implemented Clone, but it turned out to be very
confusing to people, since it created fully independent ArcSwap.
Users expected the instances to be tied to each other, that store
in one would change the result of future load of the other.
To emulate the original behaviour, one can do something like this:
let new = ArcSwap::new(old.load_full());

We do expect instances to be tied to each other, so use another Arc
over ArcSwap.

Signed-off-by: Jiang Liu <gerry@linux.alibaba.com>