stroage: fix a bug in implementation of ArcSwap::clone()
Fix a bug in implementation of ArcSwap::clone(), we should share the
ArcSwap object instead of generate a new independent one.

Previous version implemented Clone, but it turned out to be very
confusing to people, since it created fully independent ArcSwap.
Users expected the instances to be tied to each other, that store
in one would change the result of future load of the other.

Signed-off-by: Jiang Liu <gerry@linux.alibaba.com>