nydusify: support using s3 as the storage backend
This patch supports using s3 as the storage backend other than registry and oss when converting images.

Through this patch, users can use the official s3 service or other s3 compatible services (for example minio, ceph s3 gateway, etc.) as the backend to store the blob layers of the converted nydus images when executing `nydusify convert`. The usage of s3 backend is very similar with the oss backend. Users should use `--backend-type` and `--backend-config-file`(or `--backend-config`) to make the nydusify choose s3 backend.

For example:

```
nydusify convert --source <source_image> --target <target_image> --backend-type s3 --backend-config-file <path_to_config_json_file>
```

The schema of the s3 config json looks like the following, and only the `bucket_name` and the `region` are mandatory:

```
{
  "access_key_id": "access_key_id",
  "access_key_secret": "access_key_secret",
  "endpoint": "http://localhost:9000",
  "bucket_name": "bucket_name",
  "region": "us-east-1",
  "object_path-prefix": "path/to/registry/"
}
```

This patch also adds a test to verify the s3 backend works well. As the nydusd doesn't support s3 backend for now, the `nydusify check` is skipped in the end of the test.

Signed-off-by: Nan Li <loheagn@icloud.com>