tests: remove stargz test from smoke
We need to adjust the stargz conversion and run implementation,
It should be similar with zran image, we need to pack compressed
blob meta and bootstrap to a inline blob.

Let's remove the stargz test case in smoke first, then add it in
snapshotter test or smoke in golang implementation in future.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>