nydusify: refactor with converter package
Refactor converter workflow using the converter package in
acceld project:

https://github.com/goharbor/acceleration-service/pull/84

This allows nydusify to reuse acceld/snapshotter codes and
to convert layers concurrently, as well as to support the
zran conversion later.

This patch causes nydusify to rely on the containerd daemon
service, we will remove this restriction in the next commit.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>