nydus-test: enhance localfs backend to support nydus 2.2
Now nydus-image and nydus only supports the combination of:
1) nydus-image --blob-id xxx --blob xxx. The blob file name and the blob
   id stored in bootstrap will both be `xxx`.
2) nydus-image --blob-dir working-dir. The blob file name and the blob
   id stored in bootstrap will both be digest of the blob.

The nydus-test works by `nydus-image --blob xxx` without `--blob-id xxx`.
So the blob file name is xxx and the blob id stored in bootstrap is
digest of the blob. This breaks the assumption of nydusd in v2.2.

So rename the blob file from `xxx` to digest of the blob.

Signed-off-by: Jiang Liu <gerry@linux.alibaba.com>