rafs: only enable digest validate based on configuration
We found that when using the "nydus-image check --bootstrap /path/to/bootstrap"
command, it takes about 15s to check a 35MB bootstrap file (rafs v5) due to the
default digest validation. This is very slow and disabling it can reduce the time to 3s.

We should only allow this option to be configurable at runtime.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>