rafs: unify error code
Based on the error definition, InvalidData is preferable than InvalidInput.

Data not valid for the operation were encountered.
Unlike [InvalidInput], this typically means that the operation parameters were valid, however the error was caused by malformed input data.
For example, a function that reads a file into a string will error with InvalidData if the file's contents are not valid UTF-8.

Signed-off-by: Liu Bo <liub.liubo@gmail.com>