[nydusify] Bump acceleration-service to build manifests with artifactType
Nydus images built using `nydusify convert --merge-platform` will now use the `artifactType` field in the index to identify the nydus manifest. This replaces the former `os.features: "nydus.remoteimage.v1"` because using it was proving to be difficult on some registries like ECR which do not accept this field.
The SOCI snapshotter actually also use this field so if any registry doesn't support `artifactType` in the future, they will have more pressure to support this.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>